execute environment refresh command when refresh button is clicked
