+++ src/vs/workbench/contrib/positronEnvironment/browser/components/actionBars.tsx
@@ -22,6 +22,7 @@ import { PositronEnvironmentServices } from 'vs/workbench/contrib/positronEnviro
 import { usePositronEnvironmentContext } from 'vs/workbench/contrib/positronEnvironment/browser/positronEnvironmentContext';
 import { showClearEnvironmentObjectsModalDialog } from 'vs/workbench/browser/positronModalDialogs/clearEnvironmentObjectsModalDialog';
 import { LanguageRuntimeSelectorMenuButton } from 'vs/workbench/contrib/positronEnvironment/browser/components/languageRuntimeSelectorMenuButton';
+import { EnvironmentRefreshAction } from 'vs/workbench/contrib/positronEnvironment/browser/positronEnvironmentActions';
 
 // Constants.
 const kSecondaryActionBarGap = 4;
@@ -88,7 +89,7 @@ export const ActionBars = (props: PropsWithChildren<ActionBarsProps>) => {
 
 	// Refresh workspace objects handler.
 	const refreshWorkspaceObjectsHandler = () => {
-		console.log('refreshWorkspaceObjectsHandler called');
+		props.commandService.executeCommand(EnvironmentRefreshAction.ID);
 	};
 
 	// If there are no language environment, return null.
