+++ src/vs/workbench/contrib/positronEnvironment/browser/positronEnvironment.contribution.ts
@@ -10,7 +10,7 @@ import { Registry } from 'vs/platform/registry/common/platform';
 import { registerIcon } from 'vs/platform/theme/common/iconRegistry';
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ViewPaneContainer } from 'vs/workbench/browser/parts/views/viewPaneContainer';
 import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { PositronEnvironmentViewPane } from 'vs/workbench/contrib/positronEnvironment/browser/positronEnvironmentView';
@@ -18,8 +18,8 @@ import { PositronEnvironmentService } from 'vs/workbench/contrib/positronEnviron
 import { IPositronEnvironmentService, POSITRON_ENVIRONMENT_VIEW_ID } from 'vs/workbench/services/positronEnvironment/common/positronEnvironment';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions, IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { ViewContainer, IViewContainersRegistry, ViewContainerLocation, Extensions as ViewContainerExtensions, IViewsRegistry } from 'vs/workbench/common/views';
-import { Action2, registerAction2 } from 'vs/platform/actions/common/actions';
-import { ILocalizedString } from 'vs/platform/action/common/action';
+import { registerAction2 } from 'vs/platform/actions/common/actions';
+import { EnvironmentRefreshAction } from 'vs/workbench/contrib/positronEnvironment/browser/positronEnvironmentActions';
 
 // Register the Positron environment service.
 registerSingleton(IPositronEnvironmentService, PositronEnvironmentService, InstantiationType.Delayed);
@@ -55,9 +55,6 @@ Registry.as<IViewsRegistry>(ViewContainerExtensions.ViewsRegistry).registerViews
 	}
 }], VIEW_CONTAINER);
 
-export const POSITRON_ENVIRONMENT_ACTION_CATEGORY = nls.localize('positronEnvironmentCategory', "Environment");
-const category: ILocalizedString = { value: POSITRON_ENVIRONMENT_ACTION_CATEGORY, original: 'Environment' };
-
 class PositronEnvironmentContribution extends Disposable implements IWorkbenchContribution {
 	constructor(
 		@IInstantiationService instantiationService: IInstantiationService,
@@ -68,33 +65,7 @@ class PositronEnvironmentContribution extends Disposable implements IWorkbenchCo
 	}
 
 	private registerActions(): void {
-		registerAction2(class extends Action2 {
-			constructor() {
-				super({
-					id: 'workbench.action.positronEnvironment.refresh',
-					title: { value: 'Refresh Environment', original: 'Refresh Environment' },
-					f1: true,
-					category,
-					description: {
-						description: 'Refresh the list of variables and values in the Environment view.',
-						args: [{
-							name: 'options',
-							schema: {
-								type: 'object'
-							}
-						}]
-					}
-				});
-			}
-
-			/**
-			 * Runs the action and refreshes the environment.
-			 *
-			 * @param accessor The service accessor.
-			 */
-			async run(accessor: ServicesAccessor) {
-			}
-		});
+		registerAction2(EnvironmentRefreshAction);
 	}
 }
